+++ CHANGELOG.md
@@ -628,6 +628,7 @@ Thanks to the people who made this release happen!
 * jj configuration can now be stored in version control. This is known as
   "managed" configuration.
   * The config file is stored at `$REPO/.config/jj/config.toml`
+  * Existing `jj config` subcommands now have the flag `--managed`
   * For security reasons, jj will prompt you when managed config is available
     to ask you whether you trust it.